Move checklist opening to solicitacoes
