+++ AppEstoque/app/src/main/java/com/example/apestoque/fragments/ComprasFragment.kt
@@ -1,23 +1,18 @@
 package com.example.apestoque.fragments
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.example.apestoque.R
 import com.example.apestoque.adapter.SolicitacaoAdapter
-import com.example.apestoque.checklist.ChecklistActivity
 import com.example.apestoque.data.NetworkModule
 import com.example.apestoque.data.Solicitacao
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -27,7 +22,6 @@ class ComprasFragment : Fragment() {
     private lateinit var swipe: SwipeRefreshLayout
     private lateinit var rv: RecyclerView
     private lateinit var tvMsg: TextView
-    private val launcher = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { carregar() }
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -61,7 +55,7 @@ class ComprasFragment : Fragment() {
                 } else {
                     tvMsg.visibility = View.GONE
                     rv.visibility = View.VISIBLE
-                    rv.adapter = SolicitacaoAdapter(pendentes) { sol -> abrirChecklist(sol) }
+                    rv.adapter = SolicitacaoAdapter(pendentes)
                 }
             } catch (e: Exception) {
                 tvMsg.text = "Erro ao carregar"
@@ -73,11 +67,4 @@ class ComprasFragment : Fragment() {
         }
     }
 
-    private fun abrirChecklist(sol: Solicitacao) {
-        val moshi = Moshi.Builder().add(KotlinJsonAdapterFactory()).build()
-        val json = moshi.adapter(Solicitacao::class.java).toJson(sol)
-        val intent = Intent(requireContext(), ChecklistActivity::class.java)
-        intent.putExtra("solicitacao", json)
-        launcher.launch(intent)
-    }
 }
